+++ src/parser/latex/tokens.jl
@@ -45,9 +45,9 @@ $(TYPEDEF)
 A `LxCom` has a similar content as a `Block`, with the addition of the definition and a vector of brace blocks.
 """
 struct LxCom <: AbstractBlock
-    ss    ::SubString       # \\command
-    lxdef ::Ref{LxDef}      # definition of the command
-    braces::Vector{OCBlock} # relevant {...} associated with the command
+    ss    ::SubString                 # \\command
+    lxdef ::Union{Nothing,Ref{LxDef}} # definition of the command
+    braces::Vector{OCBlock}           # relevant {...} with the command
 end
 LxCom(ss, def)   = LxCom(ss, def, Vector{OCBlock}())
 from(lxc::LxCom) = from(lxc.ss)
@@ -58,10 +58,23 @@ to(lxc::LxCom)   = to(lxc.ss)
 For a given `LxCom`, retrieve the definition attached to the corresponding
 `LxDef` via the reference.
 """
-getdef(lxc::LxCom)::AS = getindex(lxc.lxdef).def
+function getdef(lxc::LxCom)::Union{Nothing,AS}
+    if isnothing(lxc.lxdef)
+        return nothing
+    end
+    return getindex(lxc.lxdef).def
+end
 
 """
 For a given `LxCom`, retrieve the name of the command via the reference.
 Example: `\\cite` --> `cite`.
 """
-getname(lxc::LxCom)::String = String(getindex(lxc.lxdef).name)[2:end]
+function getname(lxc::LxCom)::String
+    if isnothing(lxc.lxdef)
+        s = String(lxc.ss)
+        j = findfirst('{', s)
+        return lxc.ss[2:prevind(s, j)]
+    end
+    return String(getindex(lxc.lxdef).name)[2:end]
+end
+# XXX if nothing lxc.lxdef then extract from lxc.ss but extract before {}