+++ test/runtests.jl
@@ -7,9 +7,11 @@ F.FD_ENV[:SILENT_MODE] = true
 F.FD_ENV[:STRUCTURE] = v"0.1" # legacy, it's switched up in the tests.
 
 # UTILS
-println("UTILS")
+println("UTILS-1")
 include("utils/folder_structure.jl")
 include("utils/paths_vars.jl"); include("test_utils.jl")
+# ---
+println("UTILS-2")
 include("utils/misc.jl")
 include("utils/errors.jl")
 include("utils/html.jl")
@@ -35,6 +37,7 @@ include("parser/markdown+latex.jl")
 include("parser/markdown-extra.jl")
 include("parser/footnotes+links.jl")
 include("parser/latex++.jl")
+include("parser/indentation++.jl")
 println("🍺")
 
 # EVAL
@@ -96,6 +99,14 @@ begin
     rm(p; recursive=true, force=true)
 end
 cd(dirname(dirname(pathof(Franklin))))
+
+println("TEMPLATING")
+include("templating/for.jl")
+# XXX XXX 
+include("templating/fill.jl")
+
+println("UTILS FILE")
+include("utils_file/basic.jl")
 
 println("INTEGRATION")
 include("integration/literate.jl")